fix: toggleGroup 모듈 분리
