+++ packages/toggleGroup/src/index.ts
@@ -1,1 +1,2 @@
-export * from './toggleGroup';
+export * from './root';
+export * from './item';
