+++ packages/toggleGroup/src/root.tsx
@@ -1,12 +1,11 @@
 import * as React from 'react';
 
-import { composeEvent, Poly, useSafeContext } from '@over-ui/core';
+import { composeEvent, Poly } from '@over-ui/core';
 import { useControlled } from '@over-ui/use-controlled';
 import { mergeRefs } from '@over-ui/merge-refs';
-import { Toggle, ToggleProps } from '@over-ui/toggle';
 import { createRovingContext } from '@over-ui/create-roving-context';
 
-const { RovingProvider, useRoving } = createRovingContext();
+export const { RovingProvider, useRoving } = createRovingContext();
 
 // ------------------------------------------------------------------------------
 // ToggleGroup.Root
@@ -20,7 +19,7 @@ type ToggleContextType = {
   getIsPressed(value: string): boolean;
 };
 
-const ToggleContext = React.createContext<ToggleContextType | null>(null);
+export const ToggleContext = React.createContext<ToggleContextType | null>(null);
 ToggleContext.displayName = 'ToggleContext';
 
 type RootProps<T> = {
@@ -197,68 +196,6 @@ const Container: ContainerComponent = React.forwardRef(
 
 Container.displayName = 'ToggleGroup.Container';
 
-// ------------------------------------------------------------------------------
-// ToggleGroup.Item
+const ToggleGroupRoot = Root;
 
-const DEFAULT_ITEM = 'button';
-const ITEM_DISPLAY_NAME = 'ToggleGroup.Item';
-
-type ItemProps = Omit<ToggleProps, 'pressed' | 'onPressedChange' | 'defaultPressed'> & {
-  /**
-   * item의 value를 지정합니다.
-   */
-  value: string;
-};
-
-const Item: Poly.Component<typeof DEFAULT_ITEM, ItemProps> = React.forwardRef(
-  <T extends React.ElementType = typeof DEFAULT_ITEM>(
-    props: Poly.Props<T, ItemProps>,
-    ref: Poly.Ref<T>
-  ) => {
-    const { children, value, disabled = false, ...restProps } = props;
-    const ItemRef = React.useRef<HTMLButtonElement>(null);
-
-    const { multiple, setValue, deleteValue, getIsPressed } = useSafeContext(
-      ToggleContext,
-      ITEM_DISPLAY_NAME
-    );
-
-    const { useRegister } = useRoving();
-
-    useRegister(value, {
-      dom: ItemRef,
-      value,
-      disabled,
-    });
-
-    const pressed = getIsPressed(value);
-    const ariaAttrs = multiple
-      ? {}
-      : {
-          role: 'radio',
-          'aria-checked': pressed,
-          'aria-pressed': undefined,
-        };
-
-    return (
-      <Toggle
-        type="button"
-        tabIndex={-1}
-        pressed={pressed}
-        onPressedChange={(pressed) => {
-          pressed ? setValue(value) : deleteValue(value);
-        }}
-        disabled={disabled}
-        ref={mergeRefs([ItemRef, ref])}
-        {...ariaAttrs}
-        {...restProps}
-      >
-        {children}
-      </Toggle>
-    );
-  }
-);
-
-Item.displayName = 'ToggleGroup.Item';
-
-export { Root, Item };
+export { ToggleGroupRoot, Root };
